Fixed: (Rarbg) Move check response by status code to parser
